submodule changes
think i figured out the cause of that xbcore error- there was something funky with the embedded framework that caused it to error out when the submodule branch was changed.  fixed by re-adding the framework in the "embed frameworks" build phase, and by making sure that the framework was listed as an explicit dependency for the inspector.